requirements.txt
pytest==7.1.2
pytest-metadata==1.10.0
selenium==4.2.0
webdriver-manager==3.7.0
allure-pytest==2.9.45
pytest-html==3.1.1